Starting to define declarations
I want the process of defining this type to shape my eventual implementation of my module type, as modules will be constructed by application of declarations to transform them.

In the process, I'm finding that I'm able to better shape and define my term and type systems as I go, it's touch and go.