Splitting off evaluation code into it's own module
mh_term, mh_relation and mh_proposition are crowded enough as it is simply defining the form of those types. While it would be more 'pure' to define their evaluation in their seperate modules, I'm going to break off that into it's own module, given that I want that code together in one place.